Basic support of text format for parameters (#430)
* Text format support for "asyncpg" executor

Teach AsyncpgExecutor to use text format for some types of prepared
statement parameters (integers and floats). This can be controlled
with a new field "format" in the ConnectionArgs. Set it to "text"
to request a text format. The default is to use "binary" format.

BaseBinaryPostgreSQLTestCase supports a class variable FORMAT which
is the format passed to AsyncpgExecutors it creates. By default it
uses binary format, but you can set it to "text" per test case.

Acra CE does not use this functionality, but Acra EE does.

* Basic support of text format for parameters

It turns out that clients don't really use text formats for data types
which are interesting to Acra CE's transparent encryption. It works only
with byte arrays which are typically using binary format.

Text format may be occasionally used for types interesting to Acra EE:
integers and floats. However, Acra EE's encryptors *expect* such data in
text format in the first place.

Therefore, for text format of parameters we need to pass them as is.
It's the binary format which requires more attention. If the parameter
is integer -- it still must be encoded as text for the encryptor, even
if the parameter is binary.  We'll need to know what type the parameter
has and then possibly decode and reencode it after encryption. Right now
that's not possible so we continue passing binary data as is.

Passing through text parameters lets Acra EE's encryptors work correctly
when text formats are used for interesting data types. Let's add this
part at first and then get back to binary parameter recoding.

* Format constants for AsyncpgExecutor

Introduce "AsyncpgExecutor.TextFormat" and "AsyncpgExecutor.BinaryFormat"
constants instead of using hardcoded values.

* Test transparent encryption with text format

Add a test for transparent encryption with asyncpg executor using text
format where available. Acra CE does not process text format, but let's
check that it does not choke on it either.