Add certificate group generator, create two groups
Create Makefile that is able to generate pack of test certificates,
including intermediate ones. Groups are described tight there.
Add two groups: group-default that is essentially the same as
certificates in tests/ssl, and group-with-intermediate, where there is
an intermediate CA and it is used to sign acra-writer and acra-server
certificates, it also has two OCSP responder certs for two CAs.