Add CRL unit tests
Some of these functions will also be used by OCSP unit tests.
Here we have some constants that contain filenames of different
certificates that will be used during tests. Those files are decoded
from PEM format into `x509.Certificate` and then used in different
tests. These is also a helpful struct `TestCertGroup` that contains a
whole pack of certificates (and a function that generates it), some CA
and some certificates signed by this CA. Probably with an intermediate
CA. And whole certificate chains that are ready to be passed into
validators, just like TLS wrapper does.

In some stage of development, there was a Makefile that was able to
generate these groups of certificaes, it was used to create group with
intermediate CA (`tests/ssl/generate_tls_keys.sh` in its current form
cannot do this). Then, it was removed, as well as the cert group. The
tests are ready to work with 3-certificate-long chains and related code
will be uncommented when the cert group will be added (along with
something that is able to re-generate the group when needed, maybe
existing `tests/ssl/generate_tls_keys.sh` will be updated).