Perform verification of one-cert-long chains
When attempting to verify certificate chain that contains only one
certificate (that is, one self-signed certificate that belongs to list
of root certificates in TLS config), perform the verification using this
same certificate as issuer. Configured URLs won't be used in such case
as they are only appliable to leaf certificates. So only URLs from
certificate (if any) will be used (for both OCSP and CRL).

Add tests that cover such behavior. Use chain that only consists of root
CA cert and test that its verification works, as well as revocation.
Add `tests/ssl/ca/crl_with_root.pem` that contains root CA cert in list
of revoked ones (used in tests), leave a note on how to re-generate it
in `generate_tls_keys.sh`.