turn off bufferization for forked process
Our tests check that something handled correctly
according to log output.Aand some of them fork
process with log redirection to file. Before
reading the file called f.Flush to be sure
that all buffered data dumped to FS. But
some logs may be buffered in the process itself that
have written to stdout. And tests don't find
expected log entries in the output.
So now we turn off bufferization
to be sure that we have all available logs.