Test : regenerate SSL certificate for postgresql
Previous certificate didn't contain `postgresql` name in SAN, just
`localhost`:

```
openssl x509 -noout -text -in tests/ssl/postgresql/postgresql.crt

    Data:
        Version: 3 (0x2)
        Serial Number:
            3b:84:c1:a2:8a:44:d9:5f:99:ac:0b:d1:ef:19:af:36:6b:cb:b3:50
...
        Subject: C=GB, ST=London, L=London, O=Global Security, OU=IT, CN=Test leaf certificate
...
            X509v3 Subject Alternative Name:
                DNS:localhost
```

This made impossible to perform mTLS tests when Acra and PostgreSQL
are located on different hosts.

That certificate may be integrated in some pre-built Docker images, so
we intentionally do not revoke it. Here we just remove it from the repository
and regenerated with both `postgresql` and `localhost` names in SAN. So all
newer images will contain this certificate.

Also here we configured 'unique_subject = no' in tests/ssl/ca/index.txt.attr to
make it possible to not revoke previous certificate to regenerate it
with the same name.