Refactor db-connection loop to use .started
PacketHandler has a `.started` flag, which indicates whether we
process general messages (true) or expect startup packets (false)
in client connection.

In the DB-connection code we need to differentiate between these
two states too. To do so, we use firstByte variable. Therefore,
refactor this code to use `.started` flag instead of the variable.

Also, add a couple of refactoring touches here and there.