Interpret uknown startup message as an error (#511)
* Add explicit description of all supported packets

This would allow finding of new types of packets in future postgres
versions by comparing the existing table with a new one.

* Sort letters in case stmt to match the table

* Add test for SSLRequest parsing

SSLRequest consists of two parts: 4-byte length and 4-byte code.
The length is always '\x00\x00\x00\x08'. The code is always
'\x04\xd2\x16\x2f' (in decimal - [4, 210, 22, 47]).

Right now the parser checks the length only by checkign the 4th
byte, which allows passing '\x79\x08\x00\x08' as length, or any
other value, which ends with '\x08'. So, test it explicitly.

* Add simmilar test for CancelRequest

But its length is 16.

* Add test for SSLRequestError with length 16

SSLRequestError expects the length to be 8. The the parser checks
also for 16, which is a bug. Add test for that.

* Fix parser of postgres packets

The parser has a couple of bugs:

- In SSLRequest and CancelRequest it checks only a byte of length,
  instead of full length.
- SSLRequest and CancelRequest require the length of 8 and 16
  bytes respectively. But the current parser can accept SSLRequest
  with 16 bytes and CancelRequest with 8 bytes, which is not allowed.
- If the packet cannot be parsed as SSLRequest, CancelRequest or
  StartupRequest it falls to parsing it as a general packet with the
  following structure: one-byte identifier and 4-byte length. If the
  packet is garbage, it can result in a system failure due to a huge
  buffer allocation.

So fix them all by correct length check and error returning if
parsing fails.

* Fix test with unsupported packet format

This test was intended to check that we allow packets with unknown
formats. But right now we don't, so change the test.

* Add test for successful parsing of 1-byte headers

* Add test for parsing variable length messages

* Add support for GSSENCRequest

* Rename processId into processID

For the god of CI!

* Update dev-changelog

[skip ci]

* Allow unknown format as earlier

Also remove tests that don't make sense now

* Expect the first message to always be a startup

This changes the parser logic:
- If the connection is new, the first message should be one of
  startup ones. Any other message is error.
- If the startup message is received, the next messages are expected
  to be general. If format is unknown, it is parsed as general
  anyway.

* Add additional checks to tests

To check output buffer

* Update changelog

* Refactor db-connection loop to use .started

PacketHandler has a `.started` flag, which indicates whether we
process general messages (true) or expect startup packets (false)
in client connection.

In the DB-connection code we need to differentiate between these
two states too. To do so, we use firstByte variable. Therefore,
refactor this code to use `.started` flag instead of the variable.

Also, add a couple of refactoring touches here and there.

* Fix comment according to coding guidelines