Extract poison key generation interface
Right now it includes only one method:
`GeneratePoisonRecordSymmetricKey`, and it is a part of
`SymmetricEncryptionKeyStoreGenerator`. This interface in turn is
a part of:
- `StorageKeyGenerator`
- `KeyMaking`
- `ServerKeyStore`

Only `KeyMaking` uses the method for generating keys in keymaker.

So, extract separate interface dedicated solely to generation of
poison keys, in the same way as
`SymmetricEncryptionKeyStoreGenerator` is used for encryption
key generation.