Don't generate poison keys during the query (#516)
* Add .vscode into .gitignore

* Remove unused RecordProcessorKeyStore interface

Double checked that it's not used anywhere

* Regenerate with `go generate ./...`

* Extract poison key generation interface

Right now it includes only one method:
`GeneratePoisonRecordSymmetricKey`, and it is a part of
`SymmetricEncryptionKeyStoreGenerator`. This interface in turn is
a part of:
- `StorageKeyGenerator`
- `KeyMaking`
- `ServerKeyStore`

Only `KeyMaking` uses the method for generating keys in keymaker.

So, extract separate interface dedicated solely to generation of
poison keys, in the same way as
`SymmetricEncryptionKeyStoreGenerator` is used for encryption
key generation.

* Add GeneratePoisonKeyPair into PoisonKeyGenerator

And a bunch of stubs to implementors

* Rename GeneratePoison{Record,}SymmetricKey

Rename GeneratePoisonRecordSymmetricKey into
GeneratePoisonSymmetricKey to be more consistent with naming.

* Implement GeneratePoisonKeyPair for v1 keystore

* Implement GeneratePoisonKeyPair for keystore v2

* Remove PoisonKeyStore from KeyMaking interface

Also replace Get with Create in business logic code. Tests are not
touched yet.

* Remove key generation from Get* methods of v1

Get* methods generate keys, if they don't find any. So, make them
return ErrKeysNotFound instead of generating keys in keystore v1.

Also, fix tests that rely on this behaviour.

* Remove key generation from Get* methods of v2

The same as a previous commit, but for the keystore v2.

* Add warning about lack of poison keys

* Add test that poison keys don't generate on Get

* Add poison key generation tests for v1

* Add poison key tests for v2 keystore

* Add autogeneration of keys in poisonrecordmaker

And fix tests.

* Update changelog

* Fix issues for the god of CI