Fix `acra-keys list` with keystore v1 (#530)
* Fix double listing of keys in the same directory

The v1 keystore can have two paths for public and private keys.
But most often they are the same. The `ListKeys` function of keystore
doesn't check if these paths are the same, and if yes, enumerates all
keys twice.

Change this by checking the paths.

* Fix printing of log keys

The filename is `secure_log_key` and it's parsed as a regular client
key, which consists of `<client_id>_<purpose>_<kind>`. So in case of
`secure_log_key` the client_id is `secure, which is not correct.

Fix this by setting empty array as a ClientId.

* Update changelog

[skip ci]

* Move table printing code into utils

Removes code duplication.