Add response on fail for mysql (#533)
* Move EncodingValue into the base module

To reuse it later in mysql

* Move mysql.utils into base.utils

To prevent import cycles

* Add OnFail handler into text encoder

* Add OnFail handler for a binary encoding

* Merge text and binary encoding into one function

So it looks like in the postgres version. It also simplifies and
reduces the code.

* Add tests for successful, default and error paths

* Rename IdentityValue into StringValue

It makes sense, because:
a) It is used mostly in string context.
b) It is an identity operation only in postgres case. In mysql it
   encodes string as length encoded one.

* Rename ByteSequenceValue into BytesValue

To be more consistent.

* Add integration tests for `ciphertext` option

Just copy-paste from `WithoutDefaults`

* Send interrupted error with custom msg to client

In case of EncodingError, forward the it to the client as
`QueryInterruptedError` but with custom message.

* Add integration tests for response_on_fail: error

* Update changelog

* Add ValueFactory and a temporary plug

ValueFactory will be implemented by each backend to support
their own encoding.

* Implement postgresValueFactory

* Implement mysqlValueFactory

* Canonize EncodingValue and remove plugs

* Revert "Move mysql.utils into base.utils"

This reverts commit 1a913c17f02eca16f8f7025182be6656f5167a56.

Because as @lagovas mentioned:
> it's mysql related functions. they should be stored in mysql package,
  not in base)

* Continue serving packets in case of encoding error

This is temporary, until we implement proper flushing of db packets.
Right now, if data is sent after our encoding error, it would probably
trigger the "unexpected sequence number" error on the client side.