Do not abort connection after encoding error in postgresql (#537)
* Add test for simple query and transaction

Though they are failling right now.

* Add tests with prepared statements

* Add test with prepared cursors

This workflow is the most complicated, because database returns `n`
rows and then waits for another `Execute` packet. Test that
"encoding error" will not block, abort or desynchronize the connection.

* Add packet skipping in stateSkipResponse

* Update changelog

* Merge common tests between postgres and mysql

* Add explanation why we expect None after error

* Use assertRaisesRegex for checking error msg

...instead of explicitly capturing the exception and then hacking
around to find out which of the properties contains it to test that
it's the same as we expected. It also allows not caring about '\n'.

* Use new_event_loop instead of get_event_loop