Remove placeholders from connection when done (#540)
* Uncomment test lines that trigger the bug

* Rename ReadyForQueryPacket into ReadyForQuery

The name will be reused later.

* Remove placeholders on ReadyForQuery

* Process ReadyForQuery after encoding error

After the encoding error the Acra crafts and sends `Error` and
`ReadyForQuery` packets and starts skipping until database sends
`ReadyForQuery`. The latter is also skipped. This means that the Acra
doesn't proceeds ReadyForQuery in every execution path.

Change it to explicitly proceed `ReadyForQuery` from database, which
will trigger cleanup of the state.

* Update changelog