Add support for parethese expressions
Sqlalchemy serializes the test as:
```
INSERT ... VALUES (..., (_binary 'string')...)
```
So it's easier to add support also for parenthese expressions,
than fight with sqlalchemy.

The syntax in general [1] is defined as:
```
simple_expr:
  | (expr [, expr] ...)
  ...
```

But our parser supports only:
```
simple_expr:
  | (expr)
  ...
```

Which is enough for most real word cases.
Therefore, it's easy to add support for these expressions.

[1]: https://dev.mysql.com/doc/refman/8.0/en/expressions.html