Substitute oid in parse (#541)
* Add test for additional parameters in parse

This also brings new dependency - psycopg v3. This is a lot better
version of psycopg2, with new supported features, like extended
and prepared queries. In our case, we are interested in additional
type parameters in a Parse packets, which psyco specifies by default.

* Replace OID values in Parse packets

* Update changelog

* Use port var instead of hardcoded constant

Co-authored-by: Lagovas <lagovas.lagovas@gmail.com>

* Fix name of the constant in previous commit

It is DB_PORT, not TEST_DB_PORT, oops :)

* Properly convert pg placeholders into psycopg3

Replace $[0-9]+ with %s and place arguments in the correct order.

* Refactor test executors (#545)

* Extract base class from asyncpg3 mixin

* Move executors to the top

* Add asyncpg executor mixin

* Add mysql executor mixin

Co-authored-by: Lagovas <lagovas.lagovas@gmail.com>