Fix bug with i32 encoding (#542)
* Add test for additional parameters in parse

This also brings new dependency - psycopg v3. This is a lot better
version of psycopg2, with new supported features, like extended
and prepared queries. In our case, we are interested in additional
type parameters in a Parse packets, which psyco specifies by default.

* Replace OID values in Parse packets

* Update changelog

* Add test for pgBound binary int serialization

This place is responsible for normalizing values into the one form
before encrypting and putting into database. Also, this is where
the problem is.

* Fix pgBoundValue binary int normalization

By default, acra converts all integers into int64, then serializes
it into a textual representation and then encrypts.

But transforming binary int32 into int64 looked like:

  bytes -> uint32 -> int64 -> text

Which is wrong, because if in case of negative int32, the sign was
dropped at uint32 and then it wasn't restored after converting into
int64.

Fix this by properly converting int32 directly to int64, so the
compiler can insert correct instruction that can handle negative
sign:

  bytes -> uint32 -> int32 -> int64 -> text

* Handle smallint value in pgBoundValue.GetData()

This will prevent out-of-bounds panics.

* Update changelog