Add proper panic recovery
It logs the panic in the same way as the old server. The only
difference is that the new one tries to end connection with status
code 500, which I believe, is a useful addition for our users,
because such cases should not happen. But if they do, the user will
know something went terribly wrong.