Remove one goroutine from http api call
This routine caused issues with the closing the server. It awaited
either the `Done` of the context, or an error from the `http` server.
But, when the `SIGINT` or `SIGHUP` was received, the context was
closed and `runCommands` returned earlier, than the server was able
to shutdown.

Remove this goroutine. The shutdown process will be as follows:
1. Context is done, which calls `Shutdown` on the API server. When
   Shutdown is done, it calls `Close` in case of an error.
2. The server exists, which unblocks current goroutine.

If an error happens during the `Serve`, it will return with an
error. The `runCommands` will wait a timeout for other goroutines,
and close them.