[T2418] Refactor http api and bring TLS (#550)
* Add gin-based http server

* Decouple API logic from actual http routines

It allows extending and maintaining the logic separately.

* Implement /getNewZone

Generally it is a copy paste from current implementation

* Implement /resetKeyStorage

As the previous one, it's mostly a copy paste.

* Add special 404/NoRoute handler

As the old one, it should return 404 with "incorrect request".

* Add logger middleware

To output logs in our standard format with similar messages to the
old ones.

* Add proper panic recovery

It logs the panic in the same way as the old server. The only
difference is that the new one tries to end connection with status
code 500, which I believe, is a useful addition for our users,
because such cases should not happen. But if they do, the user will
know something went terribly wrong.

* Add spanning middleware

* Run the server and block until done or error

So it's done consistently as in other cases, like in `server.run`.

* Add cli flag that enables the TLS

The TLS configuration is the same as in Acra, but can easily get
its own parameters in the future if needed.

* Add warnings when TLS is not used for HTTP API

* Move translator's ConnectionToContext to network

To reuse it later in the API HTTP server.

* Use ConnectionToContextCallback in the HTTP API

As I understand, it will allow getting the client_id from the
connection, if ever needed.
Though I'm not sure how to obtain client_id the TLS is not used...

* Use gin for RunFromFileDescriptor

* Remove old http api

* Add clientID middleware

It requires setting a new config field. Maybe we should split configs
in the future - common, specifically for the acra proxy and for the
http api.

* Wrap connection into clientIDConn if no TLS

Wrap HTTP API connection into clientIDConnection if no TLS is
configured. It will allow getting the clientID in the HTTP API in
both cases: with and without TLS.

* Extract MetricConnectionCallback into network

To reuse it in the acra-server later.

* Add prometheus metrics into the http api

It requires some refactoring of the building of http api
connection wrapper, but it's ok.

* Rename client_commands into http_api

* Filter out ErrServerClosed error

It is returned after the `Shutdown` or `Close` and is doesn't mean
a problem. So, to not create false positive logs about errors,
filter it out.

* Disacrd http's server logs

They require log.Logger so there is no way of plugging the logrus.
Maybe we can provide a wrapper for that because catching logs is
useful for debug purposes.

* Add tests for HTTP API

Including tests for plain and TLS connections.

* Update changelog

[skip ci]

* Regenerate configs

* Use ioutil.Discard instead of io.Discard

Because io.Discard is not supported by the go 1.16.

* Exit when --api_use_tls but api is not enabled

Because it's invalid configuration.

* Rename AcraApiServer into HTTPAPIServer

Makes more sense and doesn't repeat itself.

* Log endpoints with info level to track who did it

* Add test for the /resetKeyStorage

* Decouple Sserver from the HTTPAPIServer

Lots of args to build the HTTPAPIServer. Maybe in the future the
builder should be introduced.

* Rename traceToLog into traceOn

Because this flag enables tracing. And when it is enabled,
the span and trace IDs are printed into the logs.

* Remove one goroutine from http api call

This routine caused issues with the closing the server. It awaited
either the `Done` of the context, or an error from the `http` server.
But, when the `SIGINT` or `SIGHUP` was received, the context was
closed and `runCommands` returned earlier, than the server was able
to shutdown.

Remove this goroutine. The shutdown process will be as follows:
1. Context is done, which calls `Shutdown` on the API server. When
   Shutdown is done, it calls `Close` in case of an error.
2. The server exists, which unblocks current goroutine.

If an error happens during the `Serve`, it will return with an
error. The `runCommands` will wait a timeout for other goroutines,
and close them.

* Fix start of the HTTP API server

It should've spawned the background routine that will stop the
server, when the context is done. But for some cases I missed that
and tests where hangling.

* Rename flag into --http_api_tls_transport_enable

To be consistent with --vault_tls_transport_enable.

* Separate HTTP logic from actual API

* Remove engine field from HTTPAPIServer

It is unused. Less code, less confusion. Less confusion, less bugs.

* Forbid empty clientID if zones are disabled

* Fix clientID check

It should be executed only when HTTP API is enabled.

* Transform error with empty clientID into warning

Because, otherwise it breaks the backward compatibility.

* Add tests for the correct client ID

They are failling for now, because there is no way of defining which
ID to use: from certificate or static.

* Allow specifying static clientID even when TLS

By saving the client id into the connection context, if
--tls_client_id_from_cert=false is provided. Not elegrant
but working solution.

* Remove ClientIDConnectionWrapper

And save the clientID into the context instead. We alredy do it,
so such change will do one abstraction less.